requirements.txt
asgiref==3.10.0
Brotli==1.1.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.2.7
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.60.1
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
